flags/types are unsigned int, the exponent shouldn't be negative and more const
git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@2040 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/libopensc/opensc.h
@@ -221,11 +221,11 @@ struct sc_pbes2_params {
 struct sc_algorithm_info {
 	unsigned int algorithm;
 	unsigned int key_length;
-	unsigned long flags;
+	unsigned int flags;
 
 	union {
 		struct sc_rsa_info {
-			long exponent;
+			unsigned long exponent;
 		} _rsa;
 	} u;
 };
